requirements.txt
Django==5.1.7
djangorestframework==3.16.0
google-cloud-translate==3.20.2
django-environ==0.12.0
langcodes==3.5.0
django-cors-headers==4.7.0